// customer-pannel/src/generated/prisma/models/Customer.ts
export type CustomerModel = runtime.Types.Result.DefaultSelection<Prisma.$CustomerPayload>
// ...
export type $CustomerPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Customer"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    customerId: string
    firstName: string | null
    lastName: string | null
    company: string | null
    city: string | null
    country: string | null
    phone1: string | null
    phone2: string | null
    email: string | null
    subscriptionDate: Date | null
    website: string | null
    aboutCustomer: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["customer"]>
  composites: {}
}